Prefer invoking preprocessor over libclang
Libclang segfaults when accessing the File.name attribute of a system
header. I was unable to determine why. Libclang also precludes showing
the full dependency graph, instead showing just the headers that were
_actually_ included.

So instead, strip out "-o" and append "-E" to the compiler arguments,
and then just _run_ it. This has some tradeoffs:

* It's slower
* You have to parse the compiler output yourself
* It requires, in the case of a Yocto eSDK toolchain, a sourced
  environment

But it also allows for more information (the full graph, cycles, etc)