Log the module, not the logger name
For each of these scripts, I just use the global logger, but in the
logs, when piping between scripts, I want to be able to identify which
script emitted each log statement.

There are two things I could do:
1. logger = logging.getLogger(name=__file__)
2. log the module name, not the logger name

I picked the second, because it was the least invasive, and easiest to
remember.